Extract Neutralinojs api calls to separate module
